requirements.txt
python-docx==0.8.11
python-dotenv==0.21.1
docx2pdf==0.1.8